+++ guestbook/urls.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 from django.conf.urls import url
-from guestbook.views import IndexView, SignView
+from guestbook.viewss import indexView,signView
 
 app_name='guestbook'
 urlpatterns = [
-    url(r'^index/$',IndexView.as_view(), name='index'),
-    url(r'^sign/$',SignView.as_view(), name='sign'),
+    url(r'^index/$',indexView.IndexView.as_view(), name='index'),
+    url(r'^sign/$',signView.SignView.as_view(), name='sign'),
 ]
