+++ templates/guestbook/main_page.html
@@ -21,6 +21,5 @@
     </form>
 
     <a href="{{ url }}">{{ url_linktext }}</a>
-
   </body>
 </html></html>
