+++ templates/guestbook/template.html
@@ -15,6 +15,8 @@
     {% endfor %}
 
 <a href="{% url sign %}?guestbook_name={{ guestbook_name }}">sign book</a>
+<a href="{{ url }}">{{ url_linktext }}</a>
 </body>
 </html>
+
